pip_requirements.txt
Django==1.9.5
django-filter==0.13.0
djangorestframework==3.3.3
pytest==2.9.1
